Now produces the DEDU model fits for V2 and V3
It’s hacky and gross. The main routine can be set to run for V2 or V3. There is a hard-coded component that forces the loading of the HRF response that was generated from V1 to be used to fit the responses for V2 and V3. The routine gets as far as saving the DEDU model fits for the 400% only stimulus runs for LMS and Mel, and then crashes as we only generated packets for these circumstances. This bit of ugly hackery was conducted to allow us to provide the reviewer with the amplitude of response within V2 and V3.

Signed-off-by: gkaguirre <gkaguirre@me.com>